services/hx-app/pyproject.toml
[tool.poetry]
name = "hx-app"
version = "0.1.0"
description = "A tournament app"
authors = ["Pedro <pedro357bm@gmail.com>"]
readme = "README.md"
packages = [{include = "hx_app"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.4"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
Markdown = "^3.4.4"
black = "^23.7.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
